+++ engine/processReports/src/main/scala/pl/touk/nussknacker/processCounts/CountsReporter.scala
@@ -1,9 +1,9 @@
 package pl.touk.nussknacker.processCounts
 
 import java.time.LocalDateTime
-
 import com.typesafe.config.Config
 
+import java.time.format.DateTimeFormatter
 import scala.concurrent.{ExecutionContext, Future}
 
 object CountsReporterCreator {
@@ -25,6 +25,15 @@ trait CountsReporter extends AutoCloseable {
 
   def prepareRawCounts(processId: String, countsRequest: CountsRequest)(implicit ec: ExecutionContext): Future[String => Option[Long]]
 
+}
+
+object CannotFetchCountsError {
+
+  private val dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")
+
+  def restartsDetected(dates: List[LocalDateTime]): CannotFetchCountsError = CannotFetchCountsError(
+    s"Counts unavailable, as process was restarted/deployed on following dates: ${dates.map(_.format(dateTimeFormatter)).mkString(", ")}"
+  )
 }
 
 case class CannotFetchCountsError(msg: String) extends Exception(msg)