+++ engine/processReports/src/test/scala/pl/touk/nussknacker/processCounts/influxdb/InfluxGeneratorSpec.scala
@@ -48,7 +48,7 @@ class InfluxGeneratorSpec extends FunSuite with Matchers with PatientScalaFuture
     |        {
     |          "name": "nodeCount",
     |          "tags": {
-    |            "action": "end",
+    |            "nodeId": "end",
     |            "slot": "1"
     |          },
     |          "columns": [
@@ -105,6 +105,6 @@ class InfluxGeneratorSpec extends FunSuite with Matchers with PatientScalaFuture
     |      ]
   """.stripMargin
 
-  val sampleInfluxOutput: List[InfluxSerie] = CirceUtil.decodeJsonUnsafe[List[InfluxSerie]](sampleInfluxOutputRaw, "failed to decode series")
+  val sampleInfluxOutput: List[InfluxSeries] = CirceUtil.decodeJsonUnsafe[List[InfluxSeries]](sampleInfluxOutputRaw, "failed to decode series")
 
 }